added the methods to save all payment info on each of the activitys
